refactor(scripts): move most of formatJSP into processJSP
`formatJSP()` is now just a thin wrapper that invokes `processJSP()` and
passes it an `onFormat` callback.

As next steps, we can add a script for linting that passes an `onLint`
callback, and a skip for generalized fixing that passes both callbacks.
